extract watcher and app
close #7
